requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.44.0
httpx==0.27.2
numpy>=2.1,<3
python-dotenv==1.0.1
